pyproject.toml
[project]
name = "twitter-forward-to-telegram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.51.0",
    "fastapi",
    "httpx",
    "openai>=1.84.0",
    "pydantic",
    "uvicorn",
    "asyncpg"
]
[dependency-groups]
alembic = [
    "alembic",
    "psycopg2-binary",
    "sqlalchemy-cockroachdb"
]
lint = [
    "ruff>=0.11.13",
]
[tool.uv]
default-groups = "all"
[tool.alembic]
script_location = "%(here)s/alembic"
prepend_sys_path = [
    ".",
    "./src"
]

# set to 'true' to search source files recursively
# in each "version_locations" directory
# new in Alembic version 1.10
# recursive_version_locations = false

# the output encoding used when revision files
# are written from script.py.mako
output_encoding = "utf-8"

# This section defines scripts or Python functions that are run
# on newly generated revision scripts.  See the documentation for further
# detail and examples
 [[tool.alembic.post_write_hooks]]
# format using "black" - use the console_scripts runner,
# against the "black" entrypoint
 name = "black"
 type = "console_scripts"
 entrypoint = "black"
 options = "-l 79 REVISION_SCRIPT_FILENAME"

 [[tool.alembic.post_write_hooks]]
# lint with attempts to fix using "ruff" - use the exec runner,
# execute a binary
 name = "ruff"
 type = "exec"
 executable = "%(here)s/.venv/bin/ruff"
 options = "check --fix REVISION_SCRIPT_FILENAME"
